FIX core : Bug on switchOn MOD web : set time for Dashboard timeout MOD utils : move dict_differ() to utils module
