+++ src/openhems/modules/util/__init__.py
@@ -12,3 +12,5 @@ from .project_configuration import ProjectConfiguration
 from .recorder import Recorder
 from .logging import getLogger, filer, get_log_file_path
 from .error import DangerousStateException
+from .utility import obj_differ
+from .json import json_default
